modeling_module/physical_problems/cellular_automaton/requirements.txt

# This is just the list of all 'pip' requirements
# Most of this is dependencies of other libraries, but
# it doesn't matter. You can write this file by yourself
# to have more control around requirements.
#
# [Hint]:
#     You no need to write exact version of library.
#     You can write just the name of library or
#     the comparable version with other version or
#     the version older/newer then some version.
#     (check the official documentation).
#
# But...
# If you are lazy (like me) you can just type
# in command line "pip freeze > requirements.txt"
# and have a nice day!

certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
matplotlib==3.3.2
moviepy==1.0.3
mpmath==1.1.0
numpy==1.19.3
Pillow==8.0.1
proglog==0.1.9
progress==1.5
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
sympy==1.6.2
tqdm==4.51.0
urllib3==1.25.11
